backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
openai>=1.40.0
pydantic>=2.0.0
python-dotenv==1.0.0
docx2pdf==0.1.8
aiofiles==23.2.1
pypdf2==3.0.1
